+++ lib/netzke/basepack/attribute_config.rb
@@ -1,5 +1,6 @@
 module Netzke
   module Basepack
+    # TODO get rid
     class AttributeConfig < ActiveSupport::OrderedOptions
       def initialize(name)
         self.name = name.to_s